@@ -37,10 +37,10 @@ setuptools.setup(
         "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
         "Operating System :: POSIX :: Linux",
         "Operating System :: Microsoft :: Windows",
-        "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
         "Programming Language :: Python :: Implementation :: CPython",
         "Topic :: Scientific/Engineering",
     ],
@@ -53,20 +53,18 @@ setuptools.setup(
     ext_modules=extension_modules,
     include_dirs=[numpy.get_include()],
     scripts=[os.path.join("hydpy", "exe", "hyd.py")],
-    python_requires=">=3.7",
+    python_requires=">=3.8",
     install_requires=[
         "black",
-        "bokeh",
         "click",
-        "Cython",
         "matplotlib",
-        "netcdf4 < 1.6; python_version == '3.7'",
-        "netcdf4; python_version > '3.7'",
+        "netcdf4",
         "networkx",
         "numpy",
         "pandas",
         "plotly",
         "scipy",
+        "setuptools",
         "typing_extensions",
         "wrapt",
         "xmlschema",
